Add custom Location entity implementation.
Support adding and removing components as well as accessing resource
identifier and filesystem path of a component through the location.

Note: At present, operations happen through the Location rather than a
collection on the Location (e.g. location['components']). In the future
the collection may be supported, but complexities need to be overcome first.

In addition, there is a need to auto-commit when adding and removing
components so that events emitted make sense and also so that data that
is managed is not accidentally orphaned.