Convert Event to subclass MutableMapping rather than dict.
Also update EventHub encoding logic in response to change.

This enables an Event to be encoded properly regardless of the depth it
appears at as the encoder is already recursive. The change to subclass
MutableMapping means that an override can be used for the encoding whereas
'dict' is handled specially by the standard JSON encoder and the encoding
logic cannot be customised.