Also detach newly created entities from Session in rollback.
A newly created entity that is not persisted should be detached in rollback as otherwise it will be cleared of all identity and
cause issues. In addition, it logically makes sense that a rolled back creation should take no further part in the session. Due
to the way Python works we cannot delete the actual object, but that may prove beneficial for debugging anyway.