Refactor optimisation to preserve integrity of database against transferred data.
By default an entry should not exist in the database until the respective data is also available. Whilst this can be
circumvented by third-parties it should be the default for the core until a change in behaviour could be
properly evaluated. This is because some behaviour and calculations rely on the availability calculations
reflecting current state rather than intent.

In addition, add more debug information for add data.