Reorder operations and add second commit to avoid bug in api backend
TODO: The commit can be avoided by reordering the operations in
the method so that the component is transferred to ftrack.server
after the thumbnail has been set.

There is currently a bug in the API backend, causing the operations
to *some* times be ordered wrongly, where the update occurs before
the component has been created, causing an integrity error.

Once this issue has been resolved, this commit can be removed and
and the update placed between component creation and registration.